Add exponential backoff policy.

+++ hekate-core/src/main/java/io/hekate/core/internal/util/ArgAssert.java
@@ -47,7 +47,7 @@ public final class ArgAssert {
         check(condition, msg);
     }
 
-    public static void positive(int val, String argName) {
+    public static void positive(long val, String argName) {
         doCheck(val > 0, argName, " must be > 0.");
     }
